Fix minor warning
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@10789 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/karts/kart_gfx.cpp
@@ -103,7 +103,7 @@ void KartGFX::addEffect(KartGFXType type, const std::string &file_name,
         kind    = NULL;
         emitter = NULL;
     }
-    assert(m_all_emitters.size()==type);
+    assert((int)m_all_emitters.size()==type);
     m_all_emitters.push_back(emitter);
     if(type==KGFX_SKID1)
         m_skid_kind1 = kind;
